refactor: 정렬 기준 Enum에 네이밍 및 패키지 위치 변경
